feat(db): embed mysql to postgres migration command
